watcher.py: Generate code with GoPy on start
Plugins utilizing Go code (e.g. WhatsApp) require bindings to be
generated via GoPy; these are currently not committed alongside other
files, and are generated on-the-fly. For local development, Slidge
expects the `generated` directory to exist, which we cannot ensure as
the `slidge` directory is mounted in from the host.

This commit ensures these auto-generated files are processed before
Slidge starts for the first time for local development.