whatsapp: Always sync contacts in FetchRoster
Fetching the WhatsApp roster will now have contact information be
propagated to Slidge at all times, even when `refresh=false`. This is to
counteract issues with missing contact information (nicknames, avatars)
when the component is added in a new client, where client-side cache is
empty.

Note that this work means that contact avatars will always be requested
from the remote, even when no refresh is requested, as there is no
server-side cache for these. In the future, one will be implemented so
as to avoid issues with rate-limiting on Slidge restart.