whatsapp: Retrieve roster after login
This commit removes the use of `FetchRoster` (which in turn would emit
`Contact` events) during login, and instead relies on pulling the roster
after login via the `Roster.fill` function, as well as with push-based
`HistorySync` and `PushName` events.

This method of roster synchonization caused a number of issues,
especially where blocking of login for a prolongued amount of time is
concerned, and moving to better supported means of synchronization is
intended with resolving these issues.

Fixes: #123