whatsapp: Add support for group-chats
WARNING: This is currently a work-in-progress and will not work.

This commit implements support for WhatsApp group-chats as XMPP MUCs,
with full support for existing features (e.g. reactions, replies, etc.)
as found in 1-to-1 chats. Future work may have us support more kinds of
groups, such as communities, etc.