core: mucs: wait for Roster.ready before returning MUC.participant_by_contact
Since the .name of the contact is used as a nickname,
we need to make sure we allowed the plugin to fill
this info, or we might end up with "unnamed" participants.
This happens for whatsapp, where incoming messages (usually
catching up missed messages since whatsmeow was down) can
arrive before the ContactInfo events.