core: send XEP-0184 delivery receipts on Contact.received()
When the "contact.received()" event does not exist on
the legacy service, we send the receipt when the legacy
API accepts the message instead.
Arguably, we could also not send receipts at all, but
I think it's nice to have a confirmation that the message
was accepted. In such legacy apps, the first check mark in
the UI corresponds to 'api accepts message'.

Fixes: https://todo.sr.ht/~nicoco/slidge/137