core: muc: mam: fix archiving of XMPP-recv messages
This fixes a subtle bug, that was only visible in
practice in a multi-device context (because an XMPP
client will rarely have to fetch the messages it has
sent itself).

It removes a useless copying of the stanza, that I
probably added because I had trouble figuring out
how to properly retrieve and set origin-id with
slixmpp.