Don't cancel Futures when waiting for Session
The `Session.wait_for_ready()` function is intended with preventing
access to incomplete sessions (e.g. during login), and is given a fairly
short timeout in order to be more responsive. However, the default
behaviour for `asyncio.wait_for()` is to cancel given `Future` instances
once the timeout has passed, which is not desirable here; fixing this
requires that we wrap these in `asyncio.shield()`.