fix: force fetch user avatar after legacy login
In theory, we receive the pubsub event at least
once from the XMPP server. In practice, this
does not always happen. It might be because we
don't store it persistently across slidge
restarts and the server only broadcast it once
and for all to us.

References: https://todo.sr.ht/~nicoco/slidge/177