perf: reduce SQL queries when broadcasting pubsub items
We only want to broadcast nick and avatar for
contacts that have been added to the roster. Up
until this commit, we "aborted" the broadcast in
the pubsub module, which called the costly
get_contact_or_group_or_participant().

This commit moves this logic from the pubsub module
to the LegacyContact class.