fix: do not fetch on-demand full VCard4 on vcard-temp request
Contact.fetch_vcard() was introduced so full VCards
do not have to be fetched systematically. This is
useful for eg, slidcord, where we don't want to fetch
profiles of all participants in a chat in advance.

A recent refactor of the VCard4 storing introduced
this undesired behaviour where we end up calling
fetch_vcard() for participants avatar requests from
the XMPP clients. This commit fixes that.