fix: ignore carbon argument when self != contact
The carbon argument should only be used for
Contacts, but we allow it for Participants too.

This commit fixes a bug where we would store group message
IDs in the direct chat XMPP msg ID->legacy msg ID store
when Participant.send_text() is called with carbon.
This also logged a useless warning.