fix: unique constraint failed with MARK_ALL_MESSAGES
With MARK_ALL_MESSAGES we keep track of msg ids
sent by each contact. This commit fixes an SQL
integrity error when a message with the same ID
is sent twice by the same contact. It is unclear
why it happens, but logging a warning is
probably better than passing the message from
the XMPP module to the XML stream.

Fixes: https://todo.sr.ht/~nicoco/slidge/213