fix: correction of messages with attachments
For compatibility with most XMPP clients, we follow the 1 message = 1
attachment rule. In this context, corrections of "a single legacy that became
several XMPP messages" are tricky to handle and mostly broken.
As a workaround, this commit treats correction of such messages by first
retracting them, then sending new ones.

Fixes: https://todo.sr.ht/~nicoco/slidgram/28