Catch IqTimeout exceptions as fatal errors
Timeouts sending `<iq>` stanzas can be quite common under contested
services, or for users with large rosters; in these cases, it's possible
to have tasks throw `IqTimeout` exceptions, which, if left uncaught, can
be fatal to the gateway as a whole.

This commit implements basic exception handling for `IqTimeout` types,
treating these as fatal errors. In the future, it might be worth
performing retries, but these will need to have an upper retry limit
applied at the least.

Closes: #222